No, I am stuck. The idea I had for the userstate is invalid. There is simply no way I will be able to do it.
https://github.com/stephan-tolksdorf/fparsec/wiki/Parsing-indentation-based-syntax-with-FParsec

And look at this stuff. The examples are written in the usual Fparsec library style, and I honestly think that while there are definite benefits to programming in this style in a language with poor optimizations, I would honestly not touch this with a 10 foot pole.

The difference with regular Fparsec combinators and this is that these require me to shuffle global state around.

That should either be done in the most effective manner or not at all.

But damn, it is 4pm today and the way it looks right now, this parser issue will sidetrack me for next few days until I internalize top down parsing or whatever I need to beat this.

I have no choice at this point.

I guess I will experience roadblocks like this every once in a while. I'll take it nice and easy. Even if it takes me a week it won't be too long in the grand scheme of things.

And when it comes to rewrite the library in Spiral, if I do it like this, I will know what I need to make it work after this. Learning excursions are inevitable in programming it seems.