Right now rather than coding, I am studying Douglas Crockford's TDOP parser example.
https://github.com/douglascrockford/TDOP

https://www.youtube.com/watch?v=9e_oEE72d3U

The Youtube video by him is quite good.

I am also reading the Javascript: The Good Parts book by him. You I'd imagine that the TDOP parsing would be immediately obvious to me given how experienced and skilled I am as a programmer, but I still absolutely beffudled.

It is always like this when studying new things. Experience and skill leave me and I am reduced to a person who never programmed before in his life. The only thing left is the desire.

That desire is the only thing I can count on. I believe in it.

I am going to read the Javascript book a bit more and then start programming for real. Mastering this technique will not be easy nor straightforward, but I am going to do it.

Looking at the source, it feels a bit inflexible. would backtracking even be possible in it? I am definitely much more comfortable with the recursive descent parsers in the Fparsec library.