The next I need the sepBy and many combinators, but I am fucking tired right now.
I had enough.

There is something I am going to add to the language.

The ErrorIfNotUnit and Ignore.

I've had the case where I forget to apply an argument sometimes and get gibbering. Since I do not have the IDE to warn me, this is a more severe error than it would be in F#.

I am going to add a special thing into the parser to insert the first Op into statements that are not at the end of the block.

That will have to wait for tomorrow.