Let me push this here for now. I need to reconsider my priorities.
Yeah, sure I can make these combinators, but I am really, really ignoring the benefits of having good error messages when I do it on my own.

As it is the Fparsec library really takes a lot of burden from my shoulders, leaving me to concentrate on the task of actually making the language.

I should probably lift the identifier and number parsers that I've made yesterday, but apart from that should leave the rest alone for the time being.

What I should be doing is really implementing Pratt parsing in the most direct possible manner in the previous version without tokenizing.

Yeah, I know that given the ridiculous amounts of backtracking the parser is doing that I would get a lot of performance benefit from that kind of thing, but error messages are really vastly more important at the current juncture.

My list of tasks going forward should be this:

1) Make the operator parser based on the previous version of the parser.

2) Go back to work on the typechecker and rewrite it completely in CPS style. This will allow me much better active patterns.

Given how focused on pattern matching the language is, that would be a definite benefit. It will also make the code a lot clearer by removing all the mutable stuff. All those stacks and sequences are something that I thought I needed, but in CPS they could be dealth away with.

3) Put in (non-recursive) algebraic datatypes into the language. This will open up a vista of new opportunities in it. There is no doubt about it, I need algebraic datatypes in some fashion if I want to do anything even remotely serious with Spiral apart from simple Cuda kernels.

Before today, I've been thinking of making a parser combinator library in it and how to write it in CPS and what would be the optimal way to do it and such.

Without algebraic datatypes it would all be impossible.

I am really most looking forward to doing #2 right now. As educational as Douglas Crockford's talks have been, I am not really interested that much in parsing. And #3 is something that I'll only consider after the typechecker is firmly done.

For the next day or two or maybe three, let me just focus on fixing the operator parser.