Before I do anything else, in order to make the modules work, I first need to rewrite expr_free_variable into CPS.
What this will allow me to do is allow me to switch to a different function and pass it the return stack if I ever hit the module case. That is the best choice.

Also, as promised, I looked at /u/combinatorylogic's nanopass stuff, but I can't really make heads or tails of it. If he had not told me that `ast.l` does that I would never have figured it out just by looking at the code.

I am kind of interested what the weak set hack for ML is now, but I'll leave that aside for now.

I am still skeptical about the Nanopass thing apart from doing it in Lisp. And I am not too great of a fan of his suggestion to do it using code generation in languages that do not have macros natively like Lisps do. And posting links to his personal code might be not be best for pedagogic purposes.

But the reason why I wanted to initiate a conversation with him is because building up everything from scratch is inspiring to me. If I had his skills, the thing I am doing right now would have been long done.

Well, let me start. Actually rather than do it in CPS, how about I accumulate names on the way forward and release them on the way down instead if I hit the CreateModule?