I fixed the modules by extending the expr_free_variables function just a tad. It works swimmingly now.
It was easier than the thought it would be and I do not need to rewrite it in CPS or anything.

I'll pause here as I have to go out again in an hour.

When I get back what I will do start work on hopefully the final typechecker version. Well, I am just kidding myself when I say that, but I do think that after I add partial evaluation and get rid of those mutable stacks that the language core will essentially be complete.

After that I will come back to where I left off, which is making the tuple zip function natively inside the language.