It is hard to believe how much a simple visit that lasts 20m can take from my day. Around 4 hours it seems.
At any rate, I needed some downtime after getting back. The car ride made me feel ill.

But now, I am ready to start.

It is time to unify all the different stacks that I am using all over the place into a single unified whole. And when I get back to the Spiral library on the ML side, I will remove that list and simply replace it with a function. That will take care of things.

But no doubt about it. Rewriting the typechecker in CPS form while being a vast improvement will also be a vast undertaking. I should aim to spread out the work over the course of a week instead of rushing it.

Looking at the expr_typecheck function, I'd be better off just erasing everything to start things off. And that is how I will start.