When I saw this infinite type error in the apply mutually recursive block, it made me groan inwardly and I had to take a break for half an hour.
I do not think any day of working on the typechecker was as much of a pain in the ass as this one.

What I will do now is start paring down the functions to the lowest possible point until I figure out how to do them right. I'll erase this here and rebuild them from the bottom up.