Given how it started, I thought it would take me a lot longer, but in the end I finished the entire typechecker today.
This is brilliant. It really did feel like it would take me something like a week, but sure enough this thing works and it works just fine. The snags I hit today and yesterday were just that. The rest neatly feel into place when I got them out of the way.

Now, converting it to CPS probably made it a lot slower compared to the previous version, but as I said, what this will allow me to do is enchance the active patterns so it will be worth it in the end.

Though, I must say, it is amazing that I managed to get rid of those recursive methods stacks. I had no idea this was possible until even recently. It is not possible to be rid of the mutable memory cell for the false case, but that is just fine. I am not trying to code in a functionaly pure manner here nor is that a virtue.

More advanced control flow in fact is, and I have it now.

It is only 5:50pm now, but I think I will call it a day here early after completing such a major undertaking.

The Spiral language has come a long way.

Looking at the 5a version of the typechecker I see that it was around 800 LOCs back then. That means that for over two month the typechecker has continually hovered at around that size as I continually refined it.

Despite the two being the same size, the difference in quality between them is trully vast.

What I will do tomorrow is slowly start work on the things outside the typechecker like the zip function and the standard library for it.

This has been a long time in coming.