1) Op Ignore and Op ErrorNonUnit. I want non-ignored non-unit expressions to be errors in Spiral. In F# they are just warnings. I'll have the Op ErrorNonUnit be inserted by the parser on non-terminal expressions. 2) I need to differentiate unsafe and safe array accesses. The later should be the default in all cases. I'll also extend Mset so it is a ternary operator and add the unsafe `<*-` as well. 3) I should extend equality so it covers tuples. Maybe extend modules so that their returnability is determined by their components. Treat the returnable modules as tuples during codegen. 4) Put in more partial evaluation into the language. After that make the push function which should be used to turn values into vars. 5) Related to #2, I need better errors. Having error propagation on par with F# or even Fparsec would be a major milestone and would significantly raised the quality of the language.
I am done with #1 for now. The transpose works, and I am ready to take the language to the next level.

I do not care about #3 and #4 right now – they won't be hard to add at any rate, but #2 and #5 which are related to errors are of utmost importance to me.

I need to take some time away from the language to research Cuda exception capabilities and in addition to that, I want to research how Fparsec and real languages in fact propagate error. Racket had that neat thing where the stack trace is graphically displayed on the IDE.

What I want is essentially a stack trace on type errors. Right now I have no idea how to do it, so I'll take a break for a few days maybe to do research.