I'll admit today was a complete waste of time, I am ashamed of myself. I wanted to study Racket and I did, but as expected I do not think I've learned anything to get a concrete benefit from it.
There are two things I've realized:

1) While the case for dynamic typing is that complete typechecking is NP Hard, Racket's macros have nothing to do with that either way. They are syntax transformers, and in some sense like Haskell's typeclasses, they are language agnostic.

Their concept is the basis for Spiral's type level functions and I will give them their proper home there.

The real benefit of functional dynamic languages lies in functions.

2) Racket's syntax is trully attrocious. Lack of polymorphism, half a dozen different lets, lack of integration of pattern matching with the aforementioned binding statements which the language designers resolved by making even more let statements.

The lack of integration between various language features is ridiculous.

Also if the conlussion is that that the true power of dynamic functional languages is in functions rather than macros, then all those parentheses are essentially just boilerplate.

---

I think I'll just stick to static functional languages thank you very much.

Spiral sort of straddles the line between dynamic and static anyway. It might be gradually typed. I am not sure how to classify it yet.

Now that I've finished the homework for the Racket course, I am through with this.

What I want is the elegance of F# with the full power of Dlang.

Today's study was literally physically painful. I could barely focus and after the first 3 hours for me and yet I did it for nearly 11. I always do these kinds of studies and I always regret it. I'll count today as a necessary sacrifice and move on.

I knew it would be like this, but I just had to do it anyway to come to a complete conclussion.

I always have the weakness to look for a better deal elsewhere, but forget it. I will never touch Lisp again. It cannot realistically compete with anything in a real world setting. F# is really the only language which speaks to my values so well and Spiral will be a marked extension to it.

It will be a DSL to trully bring out the performance aspects of F# as the metalanguage.

Apart from doing the Racket exercises, today I realized that the `Ignore` case which I added yesterday into the language can be implemented directly as `inl ignore _ = ()`. It actually did not occur to me yesterday while I was doing it. I can be foolish sometimes.

Let me just lookup region analysis for a bit and I will call it a day.

Tomorrow I will slowly start converting the typechecker with positional information and error tracing.