"1:25pm. Ok, I need a break.
Let me push this.

1:30pm. https://kseo.github.io/posts/2016-12-31-type-inference-algorithms-of-haskell-like-languages.html
https://www.microsoft.com/en-us/research/publication/outsideinx-modular-type-inference-with-local-assumptions/
OutsideIn(x) is the name of the algorithm Haskell uses for type inference.

4pm. Damn, I've been in bed for a long time thinking.
I just couldn't decide what is right.
What should be my approach to programming going forward? Is dynamic right after all to some degree.
That thing in the Spiral library where I will have to reinvent basic arithmetic, should I aim to rewrite the library on the F# side, or should I enrich my DSL and write the ML library in the Spiral language as well?
And the thing is, I am just feeling lethargic and out of energy.
Just what is expected of me here? And what should I do?
I do not feel like programming anymore for now. Now that I have these kernels, I need to decide what to do with them.
What is the correct thing to do?
Metaprogramming, it all spins around it. JDH said that the reason why F# is weak at metaprogramming is because it is weak at parsing. Spiral could be extremely strong at that.
But what about type inference? I mastered HM, but I hardly did so for the bidirectional typechecking algorithm. What I am doing now in the Spiral language can be hardly called inference. It is definitely more like abstract interpretation. And from that lens, unification makes quite a lot of sense.
Damn it, if only I could go a step forward and add metavariables to the language it would be absolutely perfect. Would that really be impossible? I do not think so.
Justice, what is justice here? Fundamentally, that argument around laziness is whether one should look at a language as an algorithmic emulator or a code generator. I am leaning towards the later as the former view is too expensive and the compilers never sufficiently smart enough, but in terms of my choice of language being F#, I picked wrong.
Ultimately, no programming language is sufficiently strong enough to be an emulator. I would not have fallen into despair 4.5 months ago and started working on Spiral if I felt that was true.
4:15pm. Control, I need more control.
I've been having a vision ever since Jdh showed me example of that parser in 13 lines.
What I could do in the Spiral language is simply insert that thing as a text string and it would all compile and interact directly in the language. This is something that would be simply unthinkable in raw F#.
4:20pm. Let me read just for a bit until my impatience and anxiety boils over.
Yeah, I am not going to be able to write the ML library in raw F#. I am in too deep for this. I just could not bear writing it in that.
Instead the path of abstract control lies in the direction of evolving my own language.

4:30pm. Let me read the Little Schemer book and then I will read the FP book. Towards the end it has a section on type inference. I also want read those papers and the Essence of ML type inference."

I am just too restless and unfocused to write any code right now. The life of a programmer is a difficult one it seems.

In the Simulacrum update I went on a rampage bashing various languages, but the outcome is not what I wanted. I wanted a concrete benefit from the language wars, but it seems to me that all of them are losers.

I did not pick wrong with F#. It is just that all of them are losers.

There is something I am certain of.

Somebody once said that the choice of languages does not matter and that good code can be written in any language.

I am absolutely, 100% against that it. My experience so far shows exactly the opposite.

The language one uses is more important than anything else in programming. It determines the ceiling and the limit of the individual.

If that was not true, I would not have gotten delayed for a year here.

I need to level up. I need to go beyond languages as they are currently offered.

I need to become the master of abstraction and controller of True Words.