"I have a lot of ideas for the language as I usually do, but in contrast to yesterday my desire is starting to crystalize.
I've made a determination to make the F# backend and to write the Spiral's ML side in the Spiral language. But despite that, or rather because I've made such a determination, I feel like I should write some useful code in Spiral first. Those Cuda kernels and layer normalization and the library.

It does not matter if it is inelegant or if I have to redo the basic arithmetic.

I'll do it properly in the end, and doing pausing the language development here in order to test out what I have and get some feedback is a viable path to take."

I did not programming today, instead I just read various computer science papers.

Tomorrow I will look at Pierce's second book. I heard about 1ML before, but its seems my education is doing its thing and I can understand something from the paper on 1ML now.

The next ML is no doubt going to be an amazing language with impredicative polymorphism, 1st class modules and the standard type inference that I know and love. If I had it, it would make the task of making the Spiral library a lot easier. I would not have to reinvent basic arithmetic for it for one. But in some other ways it would not change that much.

I still would not be able to get the exquisite control over inlining and would not get the Cuda compatible language that I want.

Nevertheless, there is a large gap between cutting edge PL research and what mainstream languages can do now.

I definitely do not feel like waiting for it to close on its own. I very well intend to jump over it.

According to this desire in me, I will spend tomorrow studying and then I will go into the testing phase for the cuda kernels.

I think I still need to add the code into the codegen for the CubBlockReduce case.

Well, I feel like I did well today. With relentleess focus, I am replenishing my motivation reserves for the next round.