Right now I am going through Pierce's ATTAPL book and the Local Type Inference paper amongst others.
The truth is, the ADTs which have abstract type parameters are throwing me for a loop. I know of a solution, but am wondering whether I could add some type inference besides at branches.

I sort of get the feeling that the CPS technique I discovered in the language is not the true extent of it.

And the way Algorithm W and bidirectional typechecking are done does not make a lot of sense to me right now. Not in the sense that I do not understand how they work, but in the sense that I am not convinced that this is their true form.

Does Haskell's OutsideIn typechecker really need 16k lines of code?

Just how would Algorithm W without metavariables work if written in CPS?

It feels so amateurish to me, the way this is presented right now. I had a lot of difficutly understanding W at first, but in order to gain true insight, I feel like I need to chisel the metavariables away.

What those things are, are control flow abstractions. The same goes for ordered contexts in the Easy Bidi paper.

I need to think about this more.

As far as the language development is concerned, I need to pause it for a bit and gather my insights and think. I'll finish Pierce's book in a day or two and then once I resume ramp down and simply get the somewhat trite stuff out of the way.

That trite stuff being:
1) Complete all the Cuda related stuff.
2) Do the Spiral library with DUs in F# as a prototype of what is to come in the completed language.

And if I still do not feel like continuing the language by then, maybe do some RL.

Whatever the case I am still yet to build interfaces to various online sites and there are loads of things I need to do in order to make my operations online.

The biggest challenge to me is definitely the making of the Spiral langauge. The ML library itself is not a big deal.

It used to be, but the typechecker I am doing right now is definitely 10x and more harder. And the Spiral language design is novel so I am definitely in the thrall of invention.