Still reading ATTAPL. To be honest even though I am only on page 162/589, I am kind of tired of the book. I'll do some more reading today and then tomorrow and then move on to writing some actual code in the new language.
Ever since I've started the new studies I've been thinking on how to do abstract datatypes, and am trying to decide what is justice.

What I trully desire is to merge ML's unification with what I have now - the way I do inference at branches is an aspect of it and I am thinking of whether it would be possible to generalize the method.

I started off by inventing the Reality algorithm, but from my current vantage point, I think it might be a part of a bigger, better thing that I cannot reach.

I feel like thinking on it for a time, while I do simpler, but necessary things. I do not want to continue work on the language until I decide what I want to do with ADTs. No doubt, I need to put them in, because ADTs are the way static languages attain flexibility of dynamic ones.

But I am not in a hurry - this language was originally developed to house Cuda kernels and I have enough for this purpose.

Even if I stopped development on it here, I now have enough to deal with the Spiral library and go into reinforcement learning.