Bad news, while I've been thinking how to do type inference for ADT generics, I've managed to prove that the Reality algorithm is wrong.
Check out the entry in the file for more details.

It took me so long to figure out, but to think in the end it was a result of myopia.

I guess, this means that adding type annotation is the first thing I am going to have to do once I resume.

Well, ignoring that, I've also figured out how to do union types and how my way of doing typechecking could lead to vast efficiencies. At first I thought I should do it like in F#, but for the non-recursive datatypes, I'd be much better off doing them in a more freeform manner.

Well, for the time being, let me continue just reading.