I thought a lot today as a read on the side.
It was a success. Today I managed to work out the semantics of union types completely, even in the recursive case and now understand how to implement it with certainty.

Well, in the Cuda backend, I would need to restrict the recursive types just like I would closures due to Cuda not having GC, nor much of a stack for that matter. I would need to restict the union types to the non-recursive case which is plenty useful on its own.

The way it will work is pretty simple and since my language already has dependent types, adding higher kinded ones will come pretty much for free.

Enough of that for today. I am quite tired now.

https://www.itu.dk/~sestoft/pebook/jonesgomardsestoft-a4.pdf

Tomorrow I will still have to decate time to reading the book linked above on partial evaluation.

But I am starting to get the feeling of wanting to get back into programming. A while ago, I was tired of it and dreaded doing those Cuda modules, but when I restart, I will get rid of that erroneous code in the If case, put in type annotations, put in Code for CubBlockReduce and then play around with those Cuda kernels a bit.

Also, I've decided to split the language into the Cuda and F# segments in the future, but I will do the utmost to ensure optimal code sharing between the two. Fundamentally, there is no point in having them as the same language as Cuda is such a fundamentally different target than .NET, but once Spiral is done, my way into RL will trully start.