"Ahhhh...I got it. and Pattern =     | A of Pattern * string // Type annotation case     | A' of Pattern * CudaTy     | S of string     | S' of string // match if not tuple     | R of Pattern list * Pattern option // Tuple     | F of Pattern * string // Functiona application with retracing.     | N of string * Pattern // matches a tuple name and proceeds onto the pattern on a hit.
and FunctionCore = string * (Pattern * Expr) list

What I will do is change so that there is a Match case in Pattern corresponding to Pattern * Expr. What that will allow me to do is also put in MatchBreak that actually passes the on_break into the Expr being evaluated much like match_f. And in fact MatchBreak and F would be pretty much identical then. I could probably reuse the code for match_f for it. Theen...

But I should probably put Pattern * Expr inside Pattern anyway. That would unify the two languagees.

What I will do is insert the break pattern into every Pattern and the call to break at the end of every typecase. The standard staments will not have them and trigger the type error, but typecases will have more specialized behavior of falling through to the next match if one of theirs fails.

Yes, this is the superior behavior, I do not need to get rid of CPS just yet it seems. And if I need things like that call/cc I will write the pattern matchers themselves in CPS and the kind of structuring I've described above will allow me to make it work."

The above is an excerp from my journal. I am done with Partial Evaluation and the Term Rewriting books. I am done not in the sense of reading them to completeness, but rather losing interest in them and concluding that there is no point in reading them through.

Partial evaluation has something to do with what I am doing now and intend to put into the language, but not really. What I am doing is really metaprogramming and not partial evaluation.

Ok, that takes care of that.

As can be seen from the excerpt today my thoughts brushed the subject of how to improve the pattern matcher.

I am going to implement fall through behavior of the kind active patterns already have into the regular typecase. What this will allow me in combination with CPS is to do arbitrary kind of metaprogramming - really powerful stuff.

Yesterday, I've also considered how I would implement things like GADTs and realized that I would have to put in boxing and a kind of dynamic typing into the langauge to make them work.

Yeah, that is not going to fly.

While reading Pierce's book he did mention that dependent typing blurs the line between compile and runtime, and I personaly would rather do without that. GADTs are definitely an example of dependent typing.

For that matter Spiral itself definitely has aspects of dependent typing, but I intend for it to only have stuff that can be resolved at compile time.

That having said, thinking about this did make me interested in how Lisps are implemented in low level languages. I am going to take some time to read Lisp in Small Pieces again. i've only brushed the book before.

Right now I am reading the book on Lua, more specifically how it is embedded in other languages. This is definitely something I am greatly interested in.

I still have some reading left to do and am not going to be able to resolve it today.

At the back of my mind, I am definitely storing up stuff to let loose once I get back.

I have a nice pace right now.

At any rate, I think the development on the Spiral language is going to continue for a long while yet, so I do not feel like rushing.

Doing it like this is best.

Mantianing a moderate stress level is important - this is one lesson I've learned in the past year.