"| ModuleT x -> Map.forall (fun _ -> get_type >> is_returnable') x
This is how module should be in is_returnable'.

I will have to make sure that the code generator handles it properly. And in destructure_deep I will have to make sure the thing works like the tuples do as well.

Also...I need to redo arrays. I have what I have right now, but I am still not happy with how it came out. It is ridiculous.

An array should be sizes and a pointer. And the reason the thing is non returnable is because pointers should be non-returnable.

Let me add one more thing to the list:
6.5) Add pointers and pointer types."

I pasted the above from my journal. Really taking a week off from coding did wonders for my brain.

I mean obviously if I am going to write level code I need pointers. Just what the hell was I doing with that StructT thing?