Done with the parser. I've done some more reading in the last two hours on combining metaprogramming and typechecking, but I really can't find anything.
As strange as it sounds, Spiral is unique.

This is both good and bad. It means I have an unique niche I can fit into and bad because I have to spend time making the tools to get into it.

Let me take a break here and I will see if the program still runs.

Before moving on to the pattern matcher, I need to make sure the code generator still works, and works with modules.

Also how are modules to be destructured, I still need to find that out.