I had a huge gust of insight right now. It should be possible to return modules from methods and methods from methods.
Previously I could not do it, because I simply did not understand renaming. Much like I could not figure out the error of the RLTY algorithm because I was too focused on the minutinae of its implementation, now that I have higher abstractions in place, they are allowing me to go further than I've gone before.

No only will I be able to do it now, but since the renamer gives me all the variables that are in scope, I will also be able to do it effectively.

Obviously, when comparing envs I will have to rename them....and the leftovers I can once more rename and collect in a tuple...

This is pretty huge. If I can do this, I will release all limits on the expressivity in this language.

This is something that I've previously thought was impossible.

Since the idea just occurred to me, I still have not worked out how to do the reasignment, but with some work it should come together.