Uah, I had so many good ideas that turned out to be wrong yesterday. What a pain in the arse.
Just what the hell was I doing for the past week?

https://www.reddit.com/r/Compilers/comments/6frlyp/for_a_functional_language_compiling_to_c_what_is/?st=j3ml8g9j&sh=6a48bf88

Also, last night I concludded that I am not even doing the if statements properly because they are not tail recursive.

And instead of putting in union types, I should just use CPS here instead. CPS can completely replace the non-recusive variants. I somehow forgot that.

Also, in the call to apply_closureT I need to rename the environments if I am passing in modules or function types. Tsk, tsk. The same goes for when doing casting.

Let me get on with it.