I am fucking pissed.
Yesterday, I had some wonderful ideas and I could almost reach out to them and touch them, but when I tried to grasp them it is like they parted like mist.

Actually, the way I would have intended to implement returns for modules and functions would have been different than before, but it is still not enough.

Aaaaagggghhhh...every single bit of complexity in my language is due to the fucking returns. From the methods and the if statements!

If only I could make the compiler go one way, everything would be 10x simpler.

I know what that is of course, it is compilation with CPS. But how do I target C with that?

I can't see it. I can't see it all.

I guess this is the limit of my ability right now.

Well, assuming that my recent changes work, the only stuff remaining would be the pattern matcher. That is definitely something I will do as it will make Spiral's metaprogramming chops trully world class.

But let me once more reiterate how frustrated I am right now. It is not just that I am compiling to C that is preventing from compiling with CPS.

The biggest problem is that I have no idea how to apply it with the way the language is currently designed. The problem is with my understanding.

Maybe I will have to look into how SML/NJ is designed. Also the whole issues with boxing is making me want to look into Lisps, but I need to focus.

Let me just finish this so I can start playing with those Cuda kernels and get back to work on the library.

Just as I expected, this language really will take me forever to master.