The typechecker is pretty much broken right now thanks to my recent changes. Let me save here.
I need to track down these bugs. Having my ideas shot down and now being mired in bugs is greatly discouraging to me.

It feel like there is so much more left to be discovered in the art of making languages and yet I can't reach it.

Right now this code look like it is dead to me, it is just pitiless, pointless complexity.

But it is at times like thse that I have to show my seriousness. I will roll back my sleeves and squish the bugs one by one.