Barring any bugs, I've put in sealing and unsealing into the typechecker. This is a pretty major addition to the language. It did not take me long to implement, but I must have spent months thinking about this.
The next step would be to add the support for it in the codegen and then I will be able to test this.

Also, since I have inlineables, methods and term functions, I think I will rename `fun` to `met` so there is less confusion with the standard ones.

Well, let me get on with it.