I put in sealing and unsealing into the codegen amongst other things. I also refactored it a little.
To be honest, I thought that putting those things, especially the sealing would be much more difficult than it was. Instead I managed to merge the code for tuples completely with the code for sealing the environment. Higher order functions can trully work miracles.

I am really looking forward to seeing how the full langauge will come out.

With this I think I am mostly set in terms of features.

The only hard part that I can envision is when it comes to doing pattern matching on recursive datatypes, is compiling the pattern matching code.

I'll deal with that when it comes to it.

Now it is time to move to testing.