Maybe I should not have implemented sealing after all. Adding that features uncovered two bugs, both of which I have no idea how to deal with.
The first one is that the typechecker will deverge on memoized calls when the inlineable is to its right.

The second that I am trying to come to grips with right now is that for some reason printing the types modifies the collection they are in, probably due to some thing with nested collections.

Modifying it inside the fold wrecks the thing.

GGGAAAaaaaa.....

Well, I'll figure out how to take care of this one, I have an idea.