I had a feeling that yesterday was a mess. I was correct. At any rate, I've figured it all out.
I've figured out how to enable optimal composability without doing any sealing whatsoever. It turns out, keeping track of arguments in the type level is completely unnecessary.

The reason why I have it because it is a throwback to where I did not know how to do proper renaming. What functions should be is much like tuples. I'll remove tags from the type level entirely.

This is it, this is why it is not good to give up and why it is good to try new features.

I'll have to do another core rewrite of the typechecker, but after I do this, the language will undergo another radical simplification.