I got the optimization pass to work.
With this the sealing feature is complete. Now modules and functions can be returned from methods, and the language is much more composable as a result.

Today was definitely beyond my expectations. I did not expect to do it this soon and this well.

The feature is less usefull than it seems, but it will make the language a lot more presentable. The most important part is that I am no longer tracking the variables at the type level. This is a huge change compared to all the previous versions of this language and a radical simplification that will pay dividents going forward.

It bears mentioning that I actually did not sacrifice anything at all by not tracking them at the type level. I was simply wrong to do this before and fixed that mistake.

I've been up for over 11h at this point and am definitely not going to be doing any more coding today.

I am not even sure what I am going to be doing tomorrow. Now that this major assault is over, I'll do a bit more slowly to get my bearings again.

The first thing that comes to mind is adding partial evaluation as the next step, but honestly, I think I would like to just write some code in the language at this point.

A bit by bit, it is definitely maturing.

It's nutrient is my desire.