Today I was busy reading papers on pattern matching. I also looked at the tutorial for the Mercury language. It reminds me of Minizinc a bit.
The pattern matching papers themselves are realyl difficult and I barely get them.

https://pdfs.semanticscholar.org/41ce/deaf715f11c3043b1e0911f8849c01ee9f68.pdf
ML pattern match compilation and partial evaluation

I am reading this one now. In the PE papers, I've seen mention a few times by now of how PE can derive the Knuth-Morris-Pratt from the naive algorithm and that is what the paper above is doing. So what I should do is before adding pattern matching is try to derive KMP myself using PE.

And if PE is capable of optimizing string matchers and pattern matchers, what will it do with parser combinators?

I was going to do parsers in Spiral, but this whole thing just made everything a lot more interesting.

I really like this approach because of the possible synergies it has with the rest of the language features.

I've been reading this stuff for the whole day now and I am starting to get tired. I'll read the above and then I will call it a day.

I will have to reread it tomorrow and hunt down some sources on how PE derives KMP before trying it myself.

Also after that I want to check out how those other languages apart from Lua do embedding. So I have a bunch of reading ahead of me.

To be honest, I had not expected that I might be able to use PE to optimize its own pattern matching.

That will definitely affect the development of the language going forward.